refactoring into hpxml 2 and 3 classes
