+++ hescorehpxml/hpxml2.py
@@ -1,0 +1,6 @@
+from .base import HPXMLtoHEScoreTranslatorBase
+
+
+class HPXML2toHEScoreTranslator(HPXMLtoHEScoreTranslatorBase):
+
+    SCHEMA_DIR = 'hpxml-2.3.0'
