+++ hescorehpxml/hpxml3.py
@@ -1,0 +1,6 @@
+from .base import HPXMLtoHEScoreTranslatorBase
+
+
+class HPXML3toHEScoreTranslator(HPXMLtoHEScoreTranslatorBase):
+
+    SCHEMA_DIR = 'hpxml-3.0.0'
