--- hescorehpxml/schemas/hpxml-1.1.1/MaintainBuilding.xsd
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns="http://hpxml.org/hpxml/2011/1"
-	targetNamespace="http://hpxml.org/hpxml/2011/1" elementFormDefault="qualified" version="1.1">
-	<xs:include schemaLocation="HPXMLDataTypes.xsd"/>
-	<xs:include schemaLocation="BaseElements.xsd"/>
-	<xs:element name="MaintainBuildingRequest">
-		<xs:complexType>
-			<xs:sequence>
-				<xs:element ref="XMLTransactionHeaderInformation"/>
-				<xs:element ref="SoftwareInfo"/>
-				<xs:element name="Building" type="Building"/>
-			</xs:sequence>
-			<xs:attribute name="schemaVersion" type="schemaVersionType"/>
-		</xs:complexType>
-	</xs:element>
-	<xs:element name="MaintainBuildingResponse">
-		<xs:complexType>
-			<xs:sequence>
-				<xs:element name="BuildingID" type="RemoteReference"/>
-				<xs:element maxOccurs="unbounded" minOccurs="0" name="StatusMessage"
-					type="StatusMessage"/>
-			</xs:sequence>
-		</xs:complexType>
-	</xs:element>
-</xs:schema>
