--- hescorehpxml/schemas/hpxml-1.1.1/MaintainConsumption.xsd
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns="http://hpxml.org/hpxml/2011/1"
-	targetNamespace="http://hpxml.org/hpxml/2011/1" elementFormDefault="qualified" version="1.1">
-	<xs:include schemaLocation="HPXMLDataTypes.xsd"/>
-	<xs:include schemaLocation="BaseElements.xsd"/>
-	<xs:element name="MaintainConsumptionRequest">
-		<xs:annotation>
-			<xs:documentation/>
-		</xs:annotation>
-		<xs:complexType>
-			<xs:sequence>
-				<xs:element ref="XMLTransactionHeaderInformation"/>
-				<xs:element ref="SoftwareInfo"/>
-				<xs:element name="Consumption" type="Consumption"/>
-			</xs:sequence>
-			<xs:attribute name="schemaVersion" type="schemaVersionType"/>
-		</xs:complexType>
-	</xs:element>
-	<xs:element name="MaintainConsumptionResponse">
-		<xs:complexType>
-			<xs:sequence>
-				<xs:element name="BuildingID" type="RemoteReference"/>
-				<xs:element maxOccurs="unbounded" minOccurs="0" name="StatusMessage"
-					type="StatusMessage"/>
-			</xs:sequence>
-		</xs:complexType>
-	</xs:element>
-</xs:schema>
