--- hescorehpxml/schemas/hpxml-1.1.1/MaintainCustomer.xsd
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- edited with XMLSpy v2011 (http://www.altova.com) by Noel Merket (NATIONAL RENEWABLE ENERGY LABORATORY) -->
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns="http://hpxml.org/hpxml/2011/1"
-	targetNamespace="http://hpxml.org/hpxml/2011/1" elementFormDefault="qualified" version="1.1">
-	<xs:include schemaLocation="HPXMLDataTypes.xsd"/>
-	<xs:include schemaLocation="BaseElements.xsd"/>
-	<xs:element name="MaintainCustomerRequest">
-		<xs:annotation>
-			<xs:documentation>Home Performance XML Root</xs:documentation>
-		</xs:annotation>
-		<xs:complexType>
-			<xs:sequence>
-				<xs:element ref="XMLTransactionHeaderInformation"/>
-				<xs:element ref="SoftwareInfo"/>
-				<xs:element name="Customer" type="Customer"/>
-			</xs:sequence>
-			<xs:attribute name="schemaVersion" type="schemaVersionType"/>
-		</xs:complexType>
-	</xs:element>
-	<xs:element name="MaintainCustomerResponse">
-		<xs:complexType>
-			<xs:sequence>
-				<xs:element name="CustomerID" type="RemoteReference"/>
-				<xs:element maxOccurs="unbounded" name="StatusMessages" type="StatusMessage"
-					minOccurs="0"/>
-			</xs:sequence>
-		</xs:complexType>
-	</xs:element>
-</xs:schema>
