--- hescorehpxml/schemas/hpxml-1.1.1/MaintainUtility.xsd
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- edited with XMLSpy v2011 (http://www.altova.com) by Noel Merket (NATIONAL RENEWABLE ENERGY LABORATORY) -->
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns="http://hpxml.org/hpxml/2011/1"
-	targetNamespace="http://hpxml.org/hpxml/2011/1" elementFormDefault="qualified" version="1.1">
-	<xs:include schemaLocation="HPXMLDataTypes.xsd"/>
-	<xs:include schemaLocation="BaseElements.xsd"/>
-	<xs:element name="MaintainUtilityRequest">
-		<xs:complexType>
-			<xs:sequence>
-				<xs:element ref="XMLTransactionHeaderInformation"/>
-				<xs:element ref="SoftwareInfo"/>
-				<xs:element name="Utility" type="Utility"/>
-			</xs:sequence>
-			<xs:attribute name="schemaVersion" type="schemaVersionType"/>
-		</xs:complexType>
-	</xs:element>
-	<xs:element name="MaintainUtilityResponse">
-		<xs:complexType>
-			<xs:sequence>
-				<xs:element maxOccurs="unbounded" name="UtilityIdentifiers" type="RemoteReference"/>
-				<xs:element maxOccurs="unbounded" minOccurs="0" name="StatusMessages"
-					type="StatusMessage"/>
-			</xs:sequence>
-		</xs:complexType>
-	</xs:element>
-</xs:schema>
